# server/swagger-spec.json
{"openapi":"3.0.0","info":{"title":"NestJs博客API","description":"博客后台管理API","version":"0.12","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"Classification":{"type":"object","properties":{"name":{"type":"string","description":"分类名称","example":"学习笔记"},"description":{"type":"string","description":"分类描述","example":"nest.js学习笔记"},"contentsNum":{"type":"number","description":"当前分类文章数目","example":"文章数"},"icon":{"type":"string","description":"分类icon","example":"https://material.iconhelper.cn/"},"rank":{"type":"number","description":"分类排序","example":""}},"required":["name","description","contentsNum","icon","rank"]},"sendCommentDto":{"type":"object","properties":{"contentsId":{"type":"string","description":"评论文章id","example":"评论文章id"},"authorId":{"type":"string","description":"评论所属内容作者id","example":"评论所属内容作者id"},"authorName":{"type":"string","description":"评论作者名称","example":"Dwsy"},"text":{"type":"string","description":"评论文字","example":"评论内容"},"ip":{"type":"string","description":"评论者ip地址","example":"233.233.233.233"},"url":{"type":"string","description":"评论者网址","example":"dwsy.link:88"},"MD5email":{"type":"string","description":"评论者邮件（加密）","example":"md5"},"email":{"type":"string","description":"评论者邮件","example":"1@1.c"},"agent":{"type":"string","description":"评论者客户端","example":"Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.96 Safari/537.36"},"isChild":{"type":"boolean","description":"isChild","example":false},"childId":{"description":"childId","example":"childId","type":"array","items":{"type":"string"}},"fatherId":{"type":"string","description":"父评论id"}},"required":["contentsId","authorId","authorName","text","ip","url","MD5email","email","agent","isChild","childId","fatherId"]},"sendChildCommentDto":{"type":"object","properties":{"contentsId":{"type":"string","description":"评论文章id","example":"评论文章id"},"authorId":{"type":"string","description":"评论所属内容作者id","example":"评论所属内容作者id"},"authorName":{"type":"string","description":"评论作者名称","example":"Dwsy"},"text":{"type":"string","description":"评论文字","example":"评论内容"},"ip":{"type":"string","description":"评论者ip地址","example":"233.233.233.233"},"url":{"type":"string","description":"评论者网址","example":"dwsy.link:88"},"MD5email":{"type":"string","description":"评论者邮件（加密）","example":"md5"},"email":{"type":"string","description":"评论者邮件","example":"1@1.c"},"agent":{"type":"string","description":"评论者客户端","example":"Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.96 Safari/537.36"},"isChild":{"type":"boolean","description":"isChild","example":false},"childId":{"description":"childId","example":"childId","type":"array","items":{"type":"string"}},"fatherId":{"type":"string","description":"父评论id"}},"required":["contentsId","authorId","authorName","text","ip","url","MD5email","email","agent","isChild","childId","fatherId"]},"Contents":{"type":"object","properties":{"fieldsId":{"type":"string","description":"文章自定义字段id","example":""},"text":{"type":"string","description":"渲染-html","example":"text文字"},"mdText":{"type":"string","description":"Markdown","example":"Markdown"},"menus":{"type":"object","description":"目录","example":"目录"},"isPublish":{"type":"boolean","description":"内容状态(是否公开)","example":true},"allowComment":{"type":"boolean","description":"是否允许评论","example":true}},"required":["fieldsId","text","mdText","menus","isPublish","allowComment"]},"ObjectId":{"type":"object","properties":{}},"Fields":{"type":"object","properties":{"contentsId":{"description":"文章id","example":"id","allOf":[{"$ref":"#/components/schemas/ObjectId"}]},"title":{"type":"string","description":"标题","example":"标题"},"cover":{"type":"string","description":"文章封面大url","example":"文章封面大"},"coverSmall":{"type":"string","description":"文章封面小url","example":"文章封面小"},"tag":{"description":"文章标签","example":"","type":"array","items":{"type":"string"}},"commentsNum":{"type":"number","description":"评论数","example":"123"},"view":{"type":"number","description":"文章浏览量","example":"123"},"classification":{"description":"文章分类 id","example":"","allOf":[{"$ref":"#/components/schemas/ObjectId"}]},"isDraft":{"type":"boolean","description":"是否为草稿","example":false}},"required":["contentsId","title","cover","coverSmall","tag","commentsNum","view","classification","isDraft"]},"PptOptions":{"type":"object","properties":{"pptImage":{"type":"string","description":"首页轮播图","example":""},"pptLink":{"type":"string","description":"首页轮播图链接","example":""},"pptTitle":{"type":"string","description":"首页轮播图标题","example":""},"pptSubTitle":{"type":"string","description":"首页轮播图副标题","example":""},"rank":{"type":"number","description":"轮播图排序","example":""}},"required":["pptImage","pptLink","pptTitle","pptSubTitle","rank"]},"contentsNum":{"type":"object","properties":{}},"Tag":{"type":"object","properties":{"name":{"type":"string","description":"标签名","example":"设计模式"},"colours":{"type":"string","description":"标签颜色","example":"primary"},"description":{"type":"string","description":"标签描述","example":"设计模式是学习心得"},"contentsNum":{"type":"number","description":"当前标签文章数目","example":"文章数"},"order":{"type":"number","description":"标签权重","example":"整数型，越高越靠前"},"icon":{"type":"string","description":"标签icon","example":"https://material.iconhelper.cn/"}},"required":["name","colours","description","contentsNum","order","icon"]},"UpFile":{"type":"object","properties":{"fileName":{"type":"string","description":"文件名","example":"xx-xxx-xxx"},"contentsId":{"description":"文件所属文章id","example":"文章id","allOf":[{"$ref":"#/components/schemas/ObjectId"}]},"url":{"type":"string","description":"文件url","example":"http://tva1.sinaimg.cn/large/005NWBIgly1gnjj15c5fsj30w20xc4gt.jpg"},"group":{"type":"string","description":"文件类型","example":"图片|压缩包|音乐|其他"}},"required":["fileName","contentsId","url","group"]},"Theme":{"type":"object","properties":{"themeName":{"type":"string","description":"主题名称","example":"light/dark"},"navigationImage":{"type":"string","description":"导航栏图片","example":""},"topImage":{"type":"string","description":"顶栏图片","example":""},"primary":{"type":"string","description":"primary","example":""},"secondary":{"type":"string","description":"secondary","example":""},"accent":{"type":"string","description":"accent","example":""},"error":{"type":"string","description":"error","example":""},"info":{"type":"string","description":"info","example":""},"success":{"type":"string","description":"success","example":""},"warning":{"type":"string","description":"warning","example":""}},"required":["themeName","navigationImage","topImage","primary","secondary","accent","error","info","success","warning"]},"Links":{"type":"object","properties":{"linksUrl":{"type":"string","description":"友链地址","example":""},"linksName":{"type":"string","description":"友链名称","example":""},"linksAvatar":{"type":"string","description":"友链头像","example":""},"linksEmail":{"type":"string","description":"友链邮箱","example":""},"linksDescription":{"type":"string","description":"友链简述","example":""},"rank":{"type":"number","description":"友链排序","example":""}},"required":["linksUrl","linksName","linksAvatar","linksEmail","linksDescription","rank"]},"RegisterDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]}}},"paths":{"/ip11111":{"get":{"operationId":"ServerController_ip","parameters":[],"responses":{"200":{"description":""}}}},"/refresh/{undefined}":{"get":{"operationId":"ServerController_refreshaCache","parameters":[{"name":"authkey","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/test":{"get":{"operationId":"ServerController_test","parameters":[{"name":"authkey","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/test1":{"get":{"operationId":"ServerController_test1","parameters":[{"name":"authkey","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/":{"get":{"operationId":"ServerController_Redirect","parameters":[],"responses":{"200":{"description":""}}}},"/api/classification/article/{id}":{"get":{"operationId":"ClassificationController_get","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["分类"]}},"/api/classification/contentsNum/{id}/{num}":{"get":{"operationId":"ClassificationController_addArticle","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"num","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["分类"],"security":[{"bearer":[]}]}},"/api/classification":{"get":{"operationId":"list","summary":"Find all records","parameters":[{"name":"query","required":false,"in":"query","description":"Query options","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["分类"]},"post":{"operationId":"ClassificationController_create","summary":"Create a record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Classification"}}}},"responses":{"201":{"description":""}},"tags":["分类"],"security":[{"bearer":[]}]}},"/api/classification/{id}":{"get":{"operationId":"ClassificationController_findOne","summary":"Find a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["分类"]},"put":{"operationId":"ClassificationController_update","summary":"Update a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Classification"}}}},"responses":{"200":{"description":""}},"tags":["分类"],"security":[{"bearer":[]}]},"delete":{"operationId":"ClassificationController_delete","summary":"Delete a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["分类"],"security":[{"bearer":[]}]}},"/api/comments":{"get":{"operationId":"CommentsController_getComments","summary":"请求all评论","parameters":[],"responses":{"200":{"description":""}},"tags":["评论"]},"post":{"operationId":"CommentsController_send","summary":"发送评论","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/sendCommentDto"}}}},"responses":{"201":{"description":""}},"tags":["评论"]}},"/api/comments/recently":{"get":{"operationId":"CommentsController_getRecentlyComments","summary":"请求最近5条评论","parameters":[],"responses":{"200":{"description":""}},"tags":["评论"]}},"/api/comments/test/{text}":{"get":{"operationId":"CommentsController_test","parameters":[{"name":"text","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["评论"]}},"/api/comments/child":{"post":{"operationId":"CommentsController_sendChild","summary":"发送子评论","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/sendChildCommentDto"}}}},"responses":{"201":{"description":""}},"tags":["评论"]}},"/api/comments/{id}":{"put":{"operationId":"CommentsController_put","summary":"修改评论","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/sendChildCommentDto"}}}},"responses":{"200":{"description":""}},"tags":["评论"],"security":[{"bearer":[]}]},"get":{"operationId":"CommentsController_get","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["评论"]}},"/api/comments/{id}/{isChild}":{"delete":{"operationId":"CommentsController_del","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"isChild","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["评论"],"security":[{"bearer":[]}]}},"/api/contents/created_toc/{id}":{"put":{"operationId":"ContentsController_createdToc","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章主体"],"security":[{"bearer":[]}]}},"/api/contents/{id}":{"get":{"operationId":"ContentsController_getContentById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章主体"]},"put":{"operationId":"ContentsController_update","summary":"Update a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Contents"}}}},"responses":{"200":{"description":""}},"tags":["文章主体"],"security":[{"bearer":[]}]},"delete":{"operationId":"ContentsController_delete","summary":"Delete a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章主体"],"security":[{"bearer":[]}]}},"/api/contents/all/{id}":{"get":{"operationId":"ContentsController_getContentById_all","summary":"页面文章请求整合","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章主体"]}},"/api/contents/search/s/{s}":{"get":{"operationId":"ContentsController_search","parameters":[{"name":"s","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章主体"]}},"/api/contents/create":{"post":{"operationId":"ContentsController_createContent","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Contents"}}}},"responses":{"201":{"description":""}},"tags":["文章主体"],"security":[{"bearer":[]}]}},"/api/contents/put/{id}":{"put":{"operationId":"ContentsController_addField","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Contents"}}}},"responses":{"200":{"description":""}},"tags":["文章主体"],"security":[{"bearer":[]}]}},"/api/contents":{"post":{"operationId":"ContentsController_create","summary":"Create a record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Contents"}}}},"responses":{"201":{"description":""}},"tags":["文章主体"],"security":[{"bearer":[]}]}},"/api/fields/archives":{"get":{"operationId":"FieldsController_archives","parameters":[],"responses":{"200":{"description":""}},"tags":["文章自定义字段"]}},"/api/fields/view/{id}":{"get":{"operationId":"FieldsController_view","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章自定义字段"]}},"/api/fields/title/{page}":{"get":{"operationId":"FieldsController_titlePage","parameters":[{"name":"page","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章自定义字段"]}},"/api/fields/tag/{id}":{"get":{"operationId":"FieldsController_getTag","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章自定义字段"]}},"/api/fields/all/{page}":{"get":{"operationId":"FieldsController_getAll","parameters":[{"name":"page","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章自定义字段"]}},"/api/fields/user":{"get":{"operationId":"FieldsController_ip","parameters":[],"responses":{"200":{"description":""}},"tags":["文章自定义字段"]}},"/api/fields/draftList":{"get":{"operationId":"list","summary":"Find all records","parameters":[],"responses":{"200":{"description":""}},"tags":["文章自定义字段"]}},"/api/fields/{id}":{"get":{"operationId":"FieldsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章自定义字段"]},"put":{"operationId":"FieldsController_update","summary":"Update a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Fields"}}}},"responses":{"200":{"description":""}},"tags":["文章自定义字段"],"security":[{"bearer":[]}]},"delete":{"operationId":"FieldsController_delete","summary":"Delete a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["文章自定义字段"],"security":[{"bearer":[]}]}},"/api/fields":{"get":{"operationId":"list","summary":"Find all records","parameters":[],"responses":{"200":{"description":""}},"tags":["文章自定义字段"]},"post":{"operationId":"FieldsController_create","summary":"Create a record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Fields"}}}},"responses":{"201":{"description":""}},"tags":["文章自定义字段"],"security":[{"bearer":[]}]}},"/api/ppt":{"get":{"operationId":"list","summary":"Find all records","parameters":[{"name":"query","required":false,"in":"query","description":"Query options","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["博客首页轮播图设置"]},"post":{"operationId":"PptController_create","summary":"Create a record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PptOptions"}}}},"responses":{"201":{"description":""}},"tags":["博客首页轮播图设置"],"security":[{"bearer":[]}]}},"/api/ppt/{id}":{"get":{"operationId":"PptController_findOne","summary":"Find a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["博客首页轮播图设置"]},"put":{"operationId":"PptController_update","summary":"Update a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PptOptions"}}}},"responses":{"200":{"description":""}},"tags":["博客首页轮播图设置"],"security":[{"bearer":[]}]},"delete":{"operationId":"PptController_delete","summary":"Delete a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["博客首页轮播图设置"],"security":[{"bearer":[]}]}},"/api/tag/article/{ids}":{"get":{"operationId":"TagController_get","parameters":[{"name":"ids","required":true,"in":"path","schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":""}},"tags":["标签"]}},"/api/tag/contentsNum/":{"post":{"operationId":"TagController_test","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/contentsNum"}}}},"responses":{"201":{"description":""}},"tags":["标签"]}},"/api/tag":{"get":{"operationId":"list","summary":"Find all records","parameters":[{"name":"query","required":false,"in":"query","description":"Query options","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["标签"]},"post":{"operationId":"TagController_create","summary":"Create a record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tag"}}}},"responses":{"201":{"description":""}},"tags":["标签"],"security":[{"bearer":[]}]}},"/api/tag/{id}":{"get":{"operationId":"TagController_findOne","summary":"Find a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["标签"]},"put":{"operationId":"TagController_update","summary":"Update a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tag"}}}},"responses":{"200":{"description":""}},"tags":["标签"],"security":[{"bearer":[]}]},"delete":{"operationId":"TagController_delete","summary":"Delete a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["标签"],"security":[{"bearer":[]}]}},"/api/upfile":{"get":{"operationId":"list","summary":"Find all records","parameters":[{"name":"query","required":false,"in":"query","description":"Query options","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["上传文件"]},"post":{"operationId":"UpfileController_create","summary":"Create a record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpFile"}}}},"responses":{"201":{"description":""}},"tags":["上传文件"],"security":[{"bearer":[]}]}},"/api/upfile/{id}":{"get":{"operationId":"UpfileController_findOne","summary":"Find a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["上传文件"]},"put":{"operationId":"UpfileController_update","summary":"Update a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpFile"}}}},"responses":{"200":{"description":""}},"tags":["上传文件"],"security":[{"bearer":[]}]},"delete":{"operationId":"UpfileController_delete","summary":"Delete a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["上传文件"],"security":[{"bearer":[]}]}},"/api/theme/pixiv":{"get":{"operationId":"ThemeController_get","parameters":[],"responses":{"200":{"description":""}},"tags":["博客主题"]}},"/api/theme/icon":{"get":{"operationId":"ThemeController_image","parameters":[],"responses":{"200":{"description":""}},"tags":["博客主题"]}},"/api/theme/tab":{"get":{"operationId":"ThemeController_tab","parameters":[],"responses":{"200":{"description":""}},"tags":["博客主题"]}},"/api/theme":{"get":{"operationId":"list","summary":"Find all records","parameters":[{"name":"query","required":false,"in":"query","description":"Query options","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["博客主题"]},"post":{"operationId":"ThemeController_create","summary":"Create a record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Theme"}}}},"responses":{"201":{"description":""}},"tags":["博客主题"],"security":[{"bearer":[]}]}},"/api/theme/{id}":{"get":{"operationId":"ThemeController_findOne","summary":"Find a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["博客主题"]},"put":{"operationId":"ThemeController_update","summary":"Update a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Theme"}}}},"responses":{"200":{"description":""}},"tags":["博客主题"],"security":[{"bearer":[]}]},"delete":{"operationId":"ThemeController_delete","summary":"Delete a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["博客主题"],"security":[{"bearer":[]}]}},"/api/links":{"get":{"operationId":"list","summary":"Find all records","parameters":[{"name":"query","required":false,"in":"query","description":"Query options","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["友情链接"]},"post":{"operationId":"LinksController_create","summary":"Create a record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Links"}}}},"responses":{"201":{"description":""}},"tags":["友情链接"],"security":[{"bearer":[]}]}},"/api/links/{id}":{"get":{"operationId":"LinksController_findOne","summary":"Find a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["友情链接"]},"put":{"operationId":"LinksController_update","summary":"Update a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Links"}}}},"responses":{"200":{"description":""}},"tags":["友情链接"],"security":[{"bearer":[]}]},"delete":{"operationId":"LinksController_delete","summary":"Delete a record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["友情链接"],"security":[{"bearer":[]}]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","summary":"注册","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["用户"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","summary":"登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["用户"]}},"/api/auth/user":{"get":{"operationId":"AuthController_user","summary":"获取个人信息","parameters":[],"responses":{"200":{"description":""}},"tags":["用户"],"security":[{"bearer":[]}]}}}}